+++ eladmin-system/src/main/java/me/zhengjie/modules/system/domain/User.java
@@ -4,12 +4,13 @@ import io.swagger.annotations.ApiModelProperty;
 import lombok.Getter;
 import lombok.Setter;
 import me.zhengjie.base.BaseEntity;
+
 import javax.persistence.*;
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 import java.io.Serializable;
-import java.util.Date;
+import java.time.LocalDateTime;
 import java.util.Objects;
 import java.util.Set;
 
@@ -84,7 +85,7 @@ public class User extends BaseEntity implements Serializable {
 
     @Column(name = "pwd_reset_time")
     @ApiModelProperty(value = "最后修改密码的时间", hidden = true)
-    private Date pwdResetTime;
+    private LocalDateTime pwdResetTime;
 
     @Override
     public boolean equals(Object o) {